+++ glide/monitors/__init__.py
@@ -1,7 +1,9 @@
+from glide.monitors.asymptotic_classical import AsymptoticClassicalMeanMonitor
 from glide.monitors.empirical_classical import EmpiricalClassicalMeanMonitor
 from glide.monitors.empirical_ppi import EmpiricalPPIMeanMonitor
 
 __all__ = [
+    "AsymptoticClassicalMeanMonitor",
     "EmpiricalClassicalMeanMonitor",
     "EmpiricalPPIMeanMonitor",
 ]